refactor(slash): unify slash and skills authority
