+++ hermes_cli/plugins.py
@@ -53,20 +53,9 @@ logger = logging.getLogger(__name__)
 def _load_quick_commands_for_slash_validation() -> dict[str, Any]:
     """Load the same effective quick-command surface slash resolution uses."""
     try:
-        from hermes_cli.config import get_hermes_home, load_config
+        from hermes_cli.commands import effective_quick_commands
 
-        config = load_config() or {}
-        if isinstance(config, dict) and config.get("quick_commands"):
-            quick = config.get("quick_commands", {})
-            return quick if isinstance(quick, dict) else {}
-
-        for project_config_path in (Path.cwd() / "cli-config.yaml", Path(__file__).resolve().parents[1] / "cli-config.yaml"):
-            if not project_config_path.exists():
-                continue
-            with open(project_config_path, encoding="utf-8") as f:
-                project_cfg = yaml.safe_load(f) or {}
-            quick = project_cfg.get("quick_commands", {}) if isinstance(project_cfg, dict) else {}
-            return quick if isinstance(quick, dict) else {}
+        return effective_quick_commands()
     except Exception:
         pass
     return {}
